+++ app/src/main/java/com/nervousfish/nervousfish/service_locator/ModuleWrapper.java
@@ -1,6 +1,6 @@
 package com.nervousfish.nervousfish.service_locator;
 
-import com.nervousfish.nervousfish.modules.IModule;
+import com.nervousfish.nervousfish.services.IModule;
 
 /**
  * Wraps a module and provides a package-private method to retrieve it so that only the service_locator classes can access the module.
