+++ app/src/main/java/com/nervousfish/nervousfish/service_locator/ServiceLocator.java
@@ -1,21 +1,21 @@
 package com.nervousfish.nervousfish.service_locator;
 
-import com.nervousfish.nervousfish.modules.constants.Constants;
-import com.nervousfish.nervousfish.modules.constants.IConstants;
-import com.nervousfish.nervousfish.modules.cryptography.EncryptorAdapter;
-import com.nervousfish.nervousfish.modules.cryptography.IEncryptor;
-import com.nervousfish.nervousfish.modules.cryptography.IKeyGenerator;
-import com.nervousfish.nervousfish.modules.cryptography.KeyGeneratorAdapter;
-import com.nervousfish.nervousfish.modules.database.GsonDatabaseAdapter;
-import com.nervousfish.nervousfish.modules.database.IDatabase;
-import com.nervousfish.nervousfish.modules.filesystem.AndroidFileSystemAdapter;
-import com.nervousfish.nervousfish.modules.filesystem.IFileSystem;
-import com.nervousfish.nervousfish.modules.pairing.DummyBluetoothHandler;
-import com.nervousfish.nervousfish.modules.pairing.DummyNFCHandler;
-import com.nervousfish.nervousfish.modules.pairing.DummyQRHandler;
-import com.nervousfish.nervousfish.modules.pairing.IBluetoothHandler;
-import com.nervousfish.nervousfish.modules.pairing.INFCHandler;
-import com.nervousfish.nervousfish.modules.pairing.IQRHandler;
+import com.nervousfish.nervousfish.services.constants.Constants;
+import com.nervousfish.nervousfish.services.constants.IConstants;
+import com.nervousfish.nervousfish.services.cryptography.EncryptorAdapter;
+import com.nervousfish.nervousfish.services.cryptography.IEncryptor;
+import com.nervousfish.nervousfish.services.cryptography.IKeyGenerator;
+import com.nervousfish.nervousfish.services.cryptography.KeyGeneratorAdapter;
+import com.nervousfish.nervousfish.services.database.GsonDatabaseAdapter;
+import com.nervousfish.nervousfish.services.database.IDatabase;
+import com.nervousfish.nervousfish.services.filesystem.AndroidFileSystemAdapter;
+import com.nervousfish.nervousfish.services.filesystem.IFileSystem;
+import com.nervousfish.nervousfish.services.pairing.DummyBluetoothHandler;
+import com.nervousfish.nervousfish.services.pairing.DummyNFCHandler;
+import com.nervousfish.nervousfish.services.pairing.DummyQRHandler;
+import com.nervousfish.nervousfish.services.pairing.IBluetoothHandler;
+import com.nervousfish.nervousfish.services.pairing.INFCHandler;
+import com.nervousfish.nervousfish.services.pairing.IQRHandler;
 
 import org.greenrobot.eventbus.EventBus;
