+++ app/src/main/java/com/nervousfish/nervousfish/services/cryptography/IEncryptor.java
@@ -1,6 +1,6 @@
-package com.nervousfish.nervousfish.modules.cryptography;
+package com.nervousfish.nervousfish.services.cryptography;
 
-import com.nervousfish.nervousfish.modules.IModule;
+import com.nervousfish.nervousfish.services.IModule;
 
 /**
  * Defines a module used for encrypting / decrypting messages that can be used by a service locator.
