+++ app/src/main/java/com/nervousfish/nervousfish/services/cryptography/IKeyGenerator.java
@@ -1,6 +1,6 @@
-package com.nervousfish.nervousfish.modules.cryptography;
+package com.nervousfish.nervousfish.services.cryptography;
 
-import com.nervousfish.nervousfish.modules.IModule;
+import com.nervousfish.nervousfish.services.IModule;
 
 /**
  * Defines a module for generating secure keys that can be used by a service locator.
