+++ app/src/main/java/com/nervousfish/nervousfish/services/filesystem/IFileSystem.java
@@ -1,6 +1,6 @@
-package com.nervousfish.nervousfish.modules.filesystem;
+package com.nervousfish.nervousfish.services.filesystem;
 
-import com.nervousfish.nervousfish.modules.IModule;
+import com.nervousfish.nervousfish.services.IModule;
 
 /**
  * Defines a module used for interacting with the filesystem that can be used by a service locator.
