+++ app/src/main/java/com/nervousfish/nervousfish/services/pairing/DummyBluetoothHandler.java
@@ -1,4 +1,4 @@
-package com.nervousfish.nervousfish.modules.pairing;
+package com.nervousfish.nervousfish.services.pairing;
 
 import com.nervousfish.nervousfish.service_locator.IServiceLocator;
 import com.nervousfish.nervousfish.service_locator.ModuleWrapper;
@@ -9,8 +9,8 @@ import org.slf4j.LoggerFactory;
 /**
  * An handler doing nothing.
  */
-public final class DummyQRHandler extends APairingHandler implements IQRHandler {
-    private static final Logger LOGGER = LoggerFactory.getLogger("DummyQRHandler");
+public final class DummyBluetoothHandler extends APairingHandler implements IBluetoothHandler {
+    private static final Logger LOGGER = LoggerFactory.getLogger("DummyBluetoothHandler");
 
     /**
      * Prevents construction from outside the class.
@@ -19,7 +19,7 @@ public final class DummyQRHandler extends APairingHandler implements IQRHandler 
      */
     @SuppressWarnings("PMD.UnusedFormalParameter")
     // This servicelocator will be used later on probably
-    private DummyQRHandler(final IServiceLocator serviceLocator) {
+    private DummyBluetoothHandler(final IServiceLocator serviceLocator) {
         super();
         LOGGER.info("Initialized");
     }
@@ -31,7 +31,7 @@ public final class DummyQRHandler extends APairingHandler implements IQRHandler 
      * @param serviceLocator The new service locator
      * @return A wrapper around a newly created instance of this class
      */
-    public static ModuleWrapper<DummyQRHandler> newInstance(final IServiceLocator serviceLocator) {
-        return new ModuleWrapper<>(new DummyQRHandler(serviceLocator));
+    public static ModuleWrapper<DummyBluetoothHandler> newInstance(final IServiceLocator serviceLocator) {
+        return new ModuleWrapper<>(new DummyBluetoothHandler(serviceLocator));
     }
 }
